backend/requirements.txt
datasets==3.0.1
Flask==3.0.3
Flask_Cors==5.0.0
fpdf==1.7.2
llama_index==0.11.14
openai==1.50.2
pandas==2.2.3
pymongo==4.9.1
pypdf==5.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.2.4
Werkzeug==3.0.4
llama-index-vector-stores-mongodb
